CSharp_Exam_1 Task_1 and Task_3 - 100%

+++ Exam_1_CSharp_Advanced/Task_2/Task_2.cs
@@ -14,8 +14,8 @@ class Task_2
 
         long[] array = Console.ReadLine().Split(',').Select(long.Parse).ToArray();
         string input;
-        BigInteger sumForward = 0;
-        BigInteger sumBackwards = 0;
+        long sumForward = 0;
+        long sumBackwards = 0;
         int position = 0;
 
         do
